Use setUrlWithoutDomain
